+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/service/DriverSupportMaintainServiceImpl.java
@@ -6,7 +6,6 @@ import com.dwarfeng.judge.stack.service.DriverSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
-import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
@@ -14,14 +13,10 @@ import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
-import com.dwarfeng.subgrade.stack.log.LogLevel;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Service
 public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainService {
@@ -29,10 +24,6 @@ public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainSe
     private final GeneralBatchCrudService<StringIdKey, DriverSupport> crudService;
     private final DaoOnlyEntireLookupService<DriverSupport> entireLookupService;
     private final DaoOnlyPresetLookupService<DriverSupport> presetLookupService;
-
-    private final List<DriverSupporter> driverSupporters;
-
-    private final ServiceExceptionMapper sem;
 
     public DriverSupportMaintainServiceImpl(
             GeneralBatchCrudService<StringIdKey, DriverSupport> crudService,
@@ -44,12 +35,6 @@ public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainSe
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;
         this.presetLookupService = presetLookupService;
-        if (Objects.isNull(driverSupporters)) {
-            this.driverSupporters = new ArrayList<>();
-        } else {
-            this.driverSupporters = driverSupporters;
-        }
-        this.sem = sem;
     }
 
     @Override
@@ -309,24 +294,5 @@ public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainSe
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public int lookupCount(String preset, Object[] objs) throws ServiceException {
         return presetLookupService.lookupCount(preset, objs);
-    }
-
-    @Override
-    @BehaviorAnalyse
-    public void reset() throws ServiceException {
-        try {
-            List<StringIdKey> driverKeys = entireLookupService.lookupAsList().stream()
-                    .map(DriverSupport::getKey).collect(Collectors.toList());
-            crudService.batchDelete(driverKeys);
-            List<DriverSupport> driverSupports = driverSupporters.stream().map(supporter -> new DriverSupport(
-                    new StringIdKey(supporter.provideType()),
-                    supporter.provideLabel(),
-                    supporter.provideDescription(),
-                    supporter.provideExampleContent()
-            )).collect(Collectors.toList());
-            crudService.batchInsert(driverSupports);
-        } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("重置触发器支持时发生异常", LogLevel.WARN, e, sem);
-        }
     }
 }